+++ libraries/session/src/main/java/androidx/media3/session/MediaControllerImplLegacy.java
@@ -621,6 +621,14 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
   }
 
   @Override
+  public ListenableFuture<SessionResult> sendCustomCommand(
+      SessionCommand command,
+      Bundle args,
+      @Nullable MediaController.ProgressListener progressListener) {
+    return sendCustomCommand(command, args);
+  }
+
+  @Override
   public Timeline getCurrentTimeline() {
     return controllerInfo.playerInfo.timeline;
   }