+++ libraries/session/src/main/java/androidx/media3/session/MediaSession.java
@@ -1448,6 +1448,29 @@ public class MediaSession {
   }
 
   /**
+   * A progress reporter to report progress for a custom command sent by a controller.
+   *
+   * <p>A non-null instance is passed to {@link MediaSession.Callback#onCustomCommand(MediaSession,
+   * ControllerInfo, SessionCommand, Bundle, ProgressReporter)} in case the controller requests
+   * progress updates.
+   */
+  @UnstableApi
+  public interface ProgressReporter {
+
+    /**
+     * Sends a progress update to the controller that has sent a custom command.
+     *
+     * <p>Updates can be sent as long as the {@link ListenableFuture<SessionCommand>} returned by
+     * {@link Callback#onCustomCommand(MediaSession, ControllerInfo, SessionCommand, Bundle,
+     * ProgressReporter)} is not done. Sending updates after completion of the future results in a
+     * no-op.
+     *
+     * @param progressData The progress data {@link Bundle} to be sent to the controller.
+     */
+    void sendProgressUpdate(Bundle progressData);
+  }
+
+  /**
    * A callback to handle incoming commands from {@link MediaController}.
    *
    * <p>The callback methods will be called from the application thread associated with the {@link
@@ -1581,6 +1604,13 @@ public class MediaSession {
      * Called when a controller sent a custom command through {@link
      * MediaController#sendCustomCommand(SessionCommand, Bundle)}.
      *
+     * <p>Note: By default {@link #onCustomCommand(MediaSession, ControllerInfo, SessionCommand,
+     * Bundle, ProgressReporter)} delegates all calls to this method and drops the option to send
+     * progress updates. If you want to implement progress updates you should override {@link
+     * #onCustomCommand(MediaSession, ControllerInfo, SessionCommand, Bundle, ProgressReporter)}
+     * instead to get access to the {@link ProgressReporter} in case a controller requests progress
+     * updates.
+     *
      * <p>{@link MediaController} instances are only allowed to send a command if the command has
      * been added to the {@link MediaSession.ConnectionResult#availableSessionCommands list of
      * available session commands} in {@link #onConnect} or set via {@link #setAvailableCommands}.
@@ -1608,6 +1638,53 @@ public class MediaSession {
         SessionCommand customCommand,
         Bundle args) {
       return Futures.immediateFuture(new SessionResult(ERROR_NOT_SUPPORTED));
+    }
+
+    /**
+     * Called when a controller sent a custom command through {@link
+     * MediaController#sendCustomCommand(SessionCommand, Bundle, MediaController.ProgressListener)}.
+     *
+     * <p>By default this callback delegates to {@link #onCustomCommand(MediaSession,
+     * ControllerInfo, SessionCommand, Bundle)}. If this method is overridden, the callback {@link
+     * #onCustomCommand(MediaSession, ControllerInfo, SessionCommand, Bundle)} is never called.
+     *
+     * <p>If a non-null {@link ProgressReporter} is passed in, then the session can report progress
+     * updates to the controller. It's the decision of the session whether or not to send progress
+     * updates. In any case, the transaction ends by completing the {@link ListenableFuture}
+     * returned by this method.
+     *
+     * <p>{@link MediaController} instances are only allowed to send a command if the command has
+     * been added to the {@link MediaSession.ConnectionResult#availableSessionCommands list of
+     * available session commands} in {@link #onConnect} or set via {@link #setAvailableCommands}.
+     *
+     * <p>Interoperability: This will be also called by {@code
+     * android.support.v4.media.MediaBrowserCompat.sendCustomAction()}. If so, {@code extras} from
+     * {@code android.support.v4.media.MediaBrowserCompat.sendCustomAction()} will be considered as
+     * {@code args} and the custom command will have {@code null} {@link
+     * SessionCommand#customExtras}.
+     *
+     * <p>Return a {@link ListenableFuture} to send a {@link SessionResult} back to the controller
+     * asynchronously. You can also return a {@link SessionResult} directly by using Guava's {@link
+     * Futures#immediateFuture(Object)}. Progress updates are dispatched only until the future has
+     * completed.
+     *
+     * @param session The session for this event.
+     * @param controller The {@linkplain ControllerInfo controller} information.
+     * @param customCommand The custom command.
+     * @param args A {@link Bundle} for additional arguments. May be empty.
+     * @param progressReporter A {@link ProgressReporter} to send progress update until the future
+     *     has completed. May be null if progress updates are not supported.
+     * @return The result of handling the custom command.
+     * @see SessionCommand#COMMAND_CODE_CUSTOM
+     */
+    @UnstableApi
+    default ListenableFuture<SessionResult> onCustomCommand(
+        MediaSession session,
+        ControllerInfo controller,
+        SessionCommand customCommand,
+        Bundle args,
+        @Nullable ProgressReporter progressReporter) {
+      return onCustomCommand(session, controller, customCommand, args);
     }
 
     /**
@@ -2201,6 +2278,9 @@ public class MediaSession {
 
     default void sendCustomCommand(int seq, SessionCommand command, Bundle args)
         throws RemoteException {}
+
+    default void sendCustomCommandProgressUpdate(
+        int seq, SessionCommand command, Bundle args, Bundle progressData) throws RemoteException {}
 
     default void onAvailableCommandsChangedFromSession(
         int seq, SessionCommands sessionCommands, Player.Commands playerCommands)